+++ packages/react-widgets/src/widgets/HistogramWidget.js
@@ -39,6 +39,7 @@ const EMPTY_ARRAY = [];
  * @param  {boolean} [props.global] - Enable/disable the viewport filtering in the data fetching.
  * @param  {Function} [props.onError] - Function to handle error messages from the widget.
  * @param  {object} [props.wrapperProps] - Extra props to pass to [WrapperWidgetUI](https://storybook-react.carto.com/?path=/docs/widgets-wrapperwidgetui--default).
+ * @param  {Function=} [props.onStateChange] - Callback to handle state updates of widgets
  * @param  {object} [props.noDataAlertProps] - Extra props to pass to [NoDataAlert]().
 
  */
@@ -61,6 +62,7 @@ function HistogramWidget({
   filterable,
   global,
   onError,
+  onStateChange,
   wrapperProps,
   noDataAlertProps
 }) {
@@ -104,6 +106,20 @@ function HistogramWidget({
     return [];
   }, [min, max, _ticks, bins, hasExternalMinMax]);
 
+  const customOnStateChange = ({ state, data }) => {
+    if (!onStateChange) return;
+    if (!data) return onStateChange({ state });
+    return onStateChange({
+      state,
+      data: {
+        values: data,
+        ticks,
+        min,
+        max
+      }
+    });
+  };
+
   let {
     data = EMPTY_ARRAY,
     isLoading,
@@ -118,6 +134,7 @@ function HistogramWidget({
     },
     global,
     onError,
+    onStateChange: customOnStateChange,
     enabled: !!ticks.length,
     attemptRemoteCalculation: _hasFeatureFlag(_FeatureFlags.REMOTE_WIDGETS)
   });