+++ CHANGELOG.md
@@ -2,6 +2,8 @@
 
 ## Not released
 
+- onStateChange callback for all widgets [#886](https://github.com/CartoDB/carto-react/pull/886)
+
 ## 3.0.0
 
 ### 3.0.0-alpha.14 (2024-06-27)